pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "nomad-tajine-plugin"
description = "An example plugin showing how NOMAD can be used outside of material science for curating food recipes."
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Hampus Näsström", email = "hampus.naesstroem@physik.hu-berlin.de" },
    { name = "FAIRmat", email = "fairmat@physik.hu-berlin.de" },
]
maintainers = [
    { name = "Hampus Näsström", email = "hampus.naesstroem@physik.hu-berlin.de" },
    { name = "FAIRmat", email = "fairmat@physik.hu-berlin.de" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.0",
    "python-magic-bin; sys_platform == 'win32'",
]

[project.urls]
Repository = "https://github.com/FAIRmat-NFDI/nomad-tajine-plugin"
Homepage = "https://fairmat-nfdi.github.io/nomad-tajine-plugin/"
"Bug Tracker" = "https://github.com/FAIRmat-NFDI/nomad-tajine-plugin/issues"
Documentation = "https://fairmat-nfdi.github.io/nomad-tajine-plugin/"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "PLC0415", # `import` should be at the top-level of a file
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]

[project.entry-points.'nomad.plugin']
parser_entry_point = "nomad_tajine_plugin.parsers:parser_entry_point"
schema_tajine_entry_point = "nomad_tajine_plugin.schema_packages:schema_tajine_entry_point"

app_entry_point = "nomad_tajine_plugin.apps:app_entry_point"
tajine_example_upload_entry_point = "nomad_tajine_plugin.example_uploads:tajine_example_upload_entry_point"
[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
